- Enforce floats in several places - Define WORD and BYTE types to SDL types - Let SDL handle some more of the win32fixes
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@739 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/keymap.c
@@ -501,14 +501,14 @@ BLOCK_HEAP  *psHeap;
 	/* Get some memory for our binding */
 	newMapping = (KEY_MAPPING*)MALLOC(sizeof(KEY_MAPPING));
 
-	ASSERT( (int)newMapping,"Couldn't allocate memory for a key mapping" );
+	ASSERT( newMapping, "Couldn't allocate memory for a key mapping" );
 
 	/* Plus one for the terminator */
 
 
 
 	newMapping->pName = (STRING*)MALLOC(strlen(name)+1);
-	ASSERT( (int)newMapping->pName,"Couldn't allocate the memory for the string in a mapping" );
+	ASSERT( newMapping->pName, "Couldn't allocate the memory for the string in a mapping" );
 
 	memSetBlockHeap(psHeap);
